requirements.dev.txt
appnope==0.1.0
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
codecov==2.0.15
coverage==4.5.1
decorator==4.3.0
idna==2.7
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
more-itertools==4.3.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.4
Pygments==2.2.0
pytest==3.7.2
pytest-tornado==0.5.0
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
